Added user and notifications

+++ src/main/java/com/crypt/decentralert/response/AddressResponse.java
@@ -6,15 +6,15 @@ import java.io.Serializable;
 
 public class AddressResponse implements Serializable {
     static final long serialVersionUID = 1L;
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    private Long id;
+    private int id;
 
     public String getNickname() {
         return nickname;
